refactor(editor): Add the concept of EditorPanel and EditorBtn
