+++ packages/artalk-sidebar/src/sidebar-root.ts
@@ -3,7 +3,7 @@ import './style/sidebar.less'
 import Context from 'artalk/types/context'
 import * as Utils from 'artalk/src/lib/utils'
 import * as Ui from 'artalk/src/lib/ui'
-import Comment  from 'artalk/src/components/comment'
+import Comment  from 'artalk/src/comment'
 import { SiteData } from 'artalk/types/artalk-data'
 import Api from 'artalk/src/api'
