+++ packages/artalk-sidebar/src/sidebar-view.ts
@@ -1,6 +1,6 @@
 import Context from 'artalk/types/context'
 import * as Utils from 'artalk/src/lib/utils'
-import Comment from 'artalk/src/components/comment'
+import Comment from 'artalk/src/comment'
 import Component from './sidebar-component'
 import { SidebarCtx } from './main'
