--- packages/artalk/src/components/_component.ts.txt
@@ -1,10 +1,0 @@
-import Context from '@/context'
-import Component from '@/lib/component'
-import * as Utils from '@/lib/utils'
-import * as Ui from '@/lib/ui'
-
-export default class Example extends Component {
-  constructor(ctx: Context) {
-    super(ctx)
-  }
-}
