--- packages/artalk/src/components/editor-plugs/_example-plug.ts.txt
@@ -1,29 +1,0 @@
-import './_PreviewPlug/PreviewPlug.less'
-import Editor from '../Editor'
-import ArtalkContext from '~/src/ArtalkContext'
-import Utils from '~/src/utils'
-
-export default class PreviewPlug extends ArtalkContext {
-  public elem: HTMLElement
-
-  constructor (public editor: Editor) {
-    super(editor.artalk)
-
-    this.initElem()
-  }
-
-  initElem () {
-    this.elem = Utils.createElement('<div class="atk-editor-plug-example"></div>')
-  }
-
-  static Name = 'example'
-  static BtnHTML = '栗子'
-
-  getEl () {
-    return this.elem
-  }
-
-  onShow () {}
-
-  onHide () {}
-}
