--- packages/artalk/src/components/editor-plugs/editor-plug.ts
@@ -1,20 +1,0 @@
-import Context from '~/types/context'
-import Editor from "../editor"
-
-export default abstract class EditorPlug {
-  protected editor: Editor
-  protected ctx: Context
-  public abstract $el: HTMLElement
-
-  constructor (editor: Editor) {
-    this.editor = editor
-    this.ctx = editor.ctx
-  }
-
-  public static Name: string
-  public static BtnHTML: string
-
-  public abstract getEl(): HTMLElement
-  public abstract onShow(): void
-  public abstract onHide(): void
-}
