+++ packages/artalk/src/components/list-lite.ts
@@ -4,7 +4,7 @@ import Component from '../lib/component'
 import * as Utils from '../lib/utils'
 import * as Ui from '../lib/ui'
 import Api from '../api'
-import Comment from './comment'
+import Comment from '../comment'
 import Pagination from './pagination'
 import ReadMoreBtn from './read-more-btn'
 import * as ListNest from './list-nest'
