+++ packages/artalk/src/editor/editor.html
@@ -7,7 +7,7 @@
   <div class="atk-textarea-wrap">
     <textarea class="atk-textarea"></textarea>
   </div>
-  <div class="atk-plug-wrap" style="display: none;"></div>
+  <div class="atk-plug-panel-wrap" style="display: none;"></div>
   <div class="atk-bottom">
     <div class="atk-item atk-plug-btn-wrap"></div>
     <div class="atk-item">
