+++ packages/artalk/src/editor/index.ts
@@ -1,0 +1,3 @@
+import Editor from './editor'
+
+export default Editor
