+++ packages/artalk/src/style/editor.less
@@ -100,7 +100,7 @@
     }
   }
 
-  & > .atk-plug-wrap {
+  & > .atk-plug-panel-wrap {
     position: relative;
     height: 180px;
     width: 100%;
